refactor: reorganize error handling and routing components
- Moved error and 404 components to frontend-specific directory
- Updated import paths to use absolute imports
- Improved error and not-found page styling and functionality
- Added VSCode TypeScript configuration
