+++ src/app/(frontend)/layout.tsx
@@ -1,9 +1,9 @@
-import './styles/globals.css';
+import "@/app/(frontend)/styles/globals.css";
 
-import type { Metadata } from 'next';
-import { Inter } from 'next/font/google';
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
 
-import Navigation from '@/app/(frontend)/components/ui/navigation';
+import Navigation from "@/app/(frontend)/components/ui/navigation";
 
 const inter = Inter({ subsets: ["latin", "latin-ext"] });
